존쟈하지 않는 게시물 삭제시 오류 매세지 출력

+++ src/main/java/com/ysh/exam/jpaBoard/article/controller/articleController.java
@@ -39,8 +39,14 @@ public class articleController {
     @RequestMapping("delete")
     @ResponseBody
     public String showDelete(long id) {
+
+        if(articleRepository.existsById(id) == false){
+            return "%d번 게시물은 이미 삭제 되었거나 없는 게시물 입니다.".formatted(id);
+        }
         articleRepository.deleteById(id);
         return "%d번 게시물이 삭제 되었습니다".formatted(id);
+
+
     }
